JPA repository, dependency injection, database seeding

+++ src/main/java/com/web/backweb/entities/User.java
@@ -12,7 +12,7 @@ public class User implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
     private String email;
@@ -24,7 +24,7 @@ public class User implements Serializable {
     public User() {
     }
 
-    public User(Integer id, String name, String email, String phone, String password) {
+    public User(Long id, String name, String email, String phone, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -34,11 +34,11 @@ public class User implements Serializable {
 
     //GETS AND SETS
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
